add typography and move header footer into folders
