+++ packages/core/stories/components/Footer/Footer.stories.js
@@ -3,7 +3,7 @@ import React from 'react';
 import Footer from '@massds/mayflower-react/dist/FooterSlim';
 import '@massds/mayflower-assets/scss/03-organisms/_footer-slim.scss';
 
-import generateTitle from '../util/generateTitle';
+import generateTitle from '../../util/generateTitle';
 
 export default {
   title: generateTitle('Footer'),
