+++ packages/core/stories/components/Header/Header.stories.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { StoryPage } from '../../.storybook/preview';
+import { StoryPage } from '../../../.storybook/preview';
 import Header from '@massds/mayflower-react/dist/HeaderSlim';
 import SiteLogo from '@massds/mayflower-react/dist/SiteLogo';
 import logo from '@massds/mayflower-assets/static/images/logo/stateseal.png';
@@ -7,7 +7,7 @@ import '@massds/mayflower-assets/scss/03-organisms/_header-slim.scss';
 import '@massds/mayflower-assets/scss/01-atoms/_site-logo.scss';
 // import HeaderDocs from './Header.md';
 
-import generateTitle from '../util/generateTitle';
+import generateTitle from '../../util/generateTitle';
 
 export default {
   title: generateTitle('Header'),
