+++ packages/core/stories/styles/index.scss
@@ -9,6 +9,7 @@
 @use '~@massds/mayflower-assets/scss/00-base/elements';
 @use '@massds/mayflower-assets/scss/01-atoms/decorative-link';
 @use '@massds/mayflower-assets/scss/01-atoms/button-copy';
+@use '@massds/mayflower-react/dist/Typography';
 
 .row.ma__row--three-up {
   margin-left: -15px;
