--- packages/core/stories/tokens/Typography.stories.mdx
@@ -1,6 +1,0 @@
-import { Meta } from '@storybook/addon-docs/blocks';
-import generateTitle from '../util/generateTitle';
-
-<Meta title={generateTitle('Typography')} />
-
-# Typography
