+++ packages/core/stories/tokens/typography/typography.json
@@ -1,0 +1,27 @@
+{
+  "characters":"ABC\u010C\u0106D\u0110EFGHIJKLMNOPQRS\u0160TUVWXYZ\u017Dabc\u010D\u0107d\u0111efghijklmnopqrs\u0161tuvwxyz\u017E\u0410\u0411\u0412\u0413\u0490\u0414\u0402\u0415\u0401\u0404\u0416\u0417\u0405\u0418\u0406\u0407\u0419\u0408\u041A\u041B\u0409\u041C\u041D\u040A\u041E\u041F\u0420\u0421\u0422\u040B\u0423\u040E\u0424\u0425\u0426\u0427\u040F\u0428\u0429\u042A\u042B\u042C\u042D\u042E\u042F\u0430\u0431\u0432\u0433\u0491\u0434\u0452\u0435\u0451\u0454\u0436\u0437\u0455\u0438\u0456\u0457\u0439\u0458\u043A\u043B\u0459\u043C\u043D\u045A\u043E\u043F\u0440\u0441\u0442\u045B\u0443\u045E\u0444\u0445\u0446\u0447\u045F\u0448\u0449\u044A\u044B\u044C\u044D\u044E\u044F\u0391\u0392\u0393\u0394\u0395\u0396\u0397\u0398\u0399\u039A\u039B\u039C\u039D\u039E\u039F\u03A0\u03A1\u03A3\u03A4\u03A5\u03A6\u03A7\u03A8\u03A9\u03B1\u03B2\u03B3\u03B4\u03B5\u03B6\u03B7\u03B8\u03B9\u03BA\u03BB\u03BC\u03BD\u03BE\u03BF\u03C0\u03C1\u03C3\u03C4\u03C5\u03C6\u03C7\u03C8\u03C9\u03AC\u0386\u03AD\u0388\u03AD\u0389\u03AF\u03CA\u0390\u038A\u03CC\u038C\u03CD\u03B0\u03CB\u038E\u03AB\u038F\u0102\u00C2\u00CA\u00D4\u01A0\u01AF\u0103\u00E2\u00EA\u00F4\u01A1\u01B01234567890\u2018?\u2019\u201C!\u201D(%)[#]{@}\/&\\<-+\u00F7\u00D7=>\u00AE\u00A9$\u20AC\u00A3\u00A5\u00A2:;,.*",
+  "letters": "abcdefghijklmnopqrstuvwxyz",
+  "numbers": "0123456789",
+  "numbersReverse": "9876543210",
+  "headings": {
+    "Heading 1": 1,
+    "Heading 2": 2,
+    "Heading 3": 3,
+    "Heading 4": 4,
+    "Heading 5": 5,
+    "Heading 6": 6,
+    "Body text": "p"
+  },
+  "styles": [
+    "lighter",
+    "lighter italic",
+    "light",
+    "light italic",
+    "normal",
+    "normal italic",
+    "bold",
+    "bold italic",
+    "bolder",
+    "bolder italic"
+  ]
+}
